@@ -52,7 +52,7 @@
         "@sveltejs/adapter-auto": "^3.3.1",
         "@sveltejs/adapter-static": "^3.0.8",
         "@sveltejs/eslint-config": "^8.1.0",
-        "@sveltejs/kit": "^2.15.0",
+        "@sveltejs/kit": "^2.15.1",
         "@sveltejs/package": "^2.3.7",
         "@sveltejs/vite-plugin-svelte": "5.0.3",
         "@sveltepress/theme-default": "^5.0.5",
@@ -83,13 +83,13 @@
         "remark-code-frontmatter": "^1.0.0",
         "sass": "^1.83.0",
         "svelte-check": "^4.1.1",
-        "svelte-highlight": "^7.8.0",
+        "svelte-highlight": "^7.8.2",
         "topojson-client": "^3.1.0",
         "tslib": "^2.8.1",
-        "typedoc": "^0.27.5",
+        "typedoc": "^0.27.6",
         "typedoc-plugin-markdown": "^4.3.3",
         "typescript": "^5.7.2",
-        "vite": "^6.0.5",
+        "vite": "^6.0.6",
         "vitest": "^2.1.8"
     },
     "types": "./dist/index.d.ts",
@@ -108,9 +108,9 @@
         "d3-scale-chromatic": "^3.1.0",
         "d3-shape": "^3.2.0",
         "d3-time": "^3.1.0",
-        "es-toolkit": "^1.30.1",
+        "es-toolkit": "^1.31.0",
         "fast-equals": "^5.0.1",
         "merge-deep": "^3.0.3",
-        "svelte": "5.15.0"
+        "svelte": "5.16.0"
     }
 }
